feat(desktop): improve markdown rendering and logging
This commit includes the following changes:

1. Improve the markdown rendering in the AI chat component by using the global `$markdown.options` instead of a local `markdownOptions` object.
2. Add more specific log level filters for `zbus` and related modules in the Tauri application setup. This helps to reduce the noise in the logs for production builds.
3. Refactor the log level configuration for the `cortex_lib` module to use a more concise syntax.

These changes aim to enhance the overall user experience and improve the maintainability of the application by optimizing the markdown rendering and the logging system.
