+++ apps/desktop/src/components/ai/chat.vue
@@ -1,20 +1,7 @@
 <script setup lang="ts">
 import { Send } from "lucide-vue-next";
 import { VueMarkdownIt } from '@f3ve/vue-markdown-it';
-import highlight from 'highlight.js/lib/core';
-import typescript from 'highlight.js/lib/languages/typescript';
-import python from 'highlight.js/lib/languages/python';
-import go from 'highlight.js/lib/languages/go';
-import java from 'highlight.js/lib/languages/java';
-import rust from 'highlight.js/lib/languages/rust';
-import 'highlight.js/styles/atom-one-dark.css';
 import type { Message } from "~/types";
-
-highlight.registerLanguage('typescript', typescript);
-highlight.registerLanguage('python', python);
-highlight.registerLanguage('go', go);
-highlight.registerLanguage('java', java);
-highlight.registerLanguage('rust', rust);
 
 interface Props {
   messages: Message[];
@@ -55,28 +42,7 @@ const handleKeydown = (event: KeyboardEvent) => {
   }
 };
 
-//* NOTE: Temporal fix for a bug in production
-const markdownOptions = {
-  html: true,
-  linkify: true,
-  typographer: true,
-  breaks: true,
-  highlight: function (str: string, lang: string) {
-    if (lang && highlight.getLanguage(lang)) {
-      try {
-        const result = highlight.highlight(str, {
-          language: lang,
-          ignoreIllegals: true
-        }).value;
-        return `<pre class="hljs language-${lang}"><code>${result}</code></pre>`;
-      } catch (error) {
-        console.error('Error highlighting code:', error);
-      }
-    }
-    // Fallback to plain text if the language is not supported
-    return `<pre class="hljs"><code>${str}</code></pre>`;
-  }
-};
+const { $markdown } = useNuxtApp();
 </script>
 
 <template>
@@ -131,7 +97,7 @@ const markdownOptions = {
               <div class="prose dark:prose-invert max-w-none py-5">
                 <VueMarkdownIt
                     :source="message.content"
-                    :options="markdownOptions"
+                    :options="$markdown.options"
                     class="text-xs sm:text-sm text-foreground break-words"
                 />
               </div>
@@ -161,7 +127,7 @@ const markdownOptions = {
             <div class="prose dark:prose-invert max-w-none py-5">
               <VueMarkdownIt
                   :source="currentStreamingMessage"
-                  :options="markdownOptions"
+                  :options="$markdown.options"
                   class="text-xs sm:text-sm text-foreground break-words"
               />
             </div>
@@ -189,57 +155,3 @@ const markdownOptions = {
     </div>
   </div>
 </template>
-
-<style>
-.prose {
-  width: 100%;
-  color: inherit;
-}
-
-.prose pre {
-  padding: 1rem;
-  border-radius: 0.5rem;
-  overflow-x: auto;
-  margin: 1rem 0;
-}
-
-.prose code {
-  color: inherit;
-  padding: 0.2rem 0.4rem;
-  border-radius: 0.25rem;
-  font-size: 0.875em;
-}
-
-.prose pre code {
-  padding: 0;
-  border-radius: 0;
-  background: none;
-}
-
-/* Estilos específicos para el tema oscuro */
-.dark .prose .hljs {
-  background: #1e1e1e;
-  color: #d4d4d4;
-}
-
-/* Ajustes adicionales para highlight.js */
-.hljs {
-  background: #282c34;
-  color: #abb2bf;
-  padding: 1em;
-  border-radius: 0.5em;
-}
-
-@keyframes bounce {
-  0%, 100% {
-    transform: translateY(0);
-  }
-  50% {
-    transform: translateY(-4px);
-  }
-}
-
-.animate-bounce {
-  animation: bounce 0.6s infinite;
-}
-</style>
